rallocated searchbar as new component
